+++ hoomd/filter/rigid.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
+
 """Define the Rigid filter."""
 
 from hoomd.filter.filter_ import ParticleFilter