+++ hoomd/logging.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2009-2021 The Regents of the University of Michigan
-# This file is part of the HOOMD-blue project, released under the BSD 3-Clause
-# License.
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 """Logging infrastructure."""
 
@@ -48,8 +47,6 @@ class LoggerCategories(Flag):
 
         particle: per-particle quantity
 
-        state: internal category for specifying object's internal state
-
         ALL: a combination of all other categories
 
         NONE: represents no category
@@ -67,7 +64,6 @@ class LoggerCategories(Flag):
     improper = auto()
     pair = auto()
     particle = auto()
-    state = auto()
 
     @classmethod
     def any(cls, categories=None):
@@ -490,8 +486,6 @@ class _LoggerEntry:
         except DataAccessError:
             attr = None
 
-        if self.category is LoggerCategories.state:
-            return attr
         if callable(attr):
             return (attr(), self.category.name)
         else: