+++ hoomd/md/pair/pair.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2009-2021 The Regents of the University of Michigan
-# This file is part of the HOOMD-blue project, released under the BSD 3-Clause
-# License.
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 """Pair potentials."""
 
@@ -449,8 +448,76 @@ class SLJ(Pair):
         # this potential needs diameter shifting on
         self._nlist.diameter_shift = True
 
-        # NOTE do we need something to automatically set the max_diameter
-        # correctly?
+
+class ExpandedLJ(Pair):
+    r"""Expanded Lennard-Jones pair potential.
+
+    Args:
+        nlist (`hoomd.md.nlist.NList`): Neighbor list.
+        default_r_cut (float): Default cutoff radius :math:`[\mathrm{length}]`.
+        default_r_on (float): Default turn-on radius :math:`[\mathrm{length}]`.
+        mode (str): Energy shifting mode.
+
+    `ExpandedLJ` specifies that a radially-shifted Lennard-Jones
+    type pair potential should be applied between every non-excluded
+    particle pair in the simulation.
+
+    .. math::
+        :nowrap:
+
+        \begin{eqnarray*}
+        V_{\mathrm{ExpandedLJ}}(r)  = & 4 \varepsilon \left[ \left(
+                                \frac{\sigma}{r - \Delta} \right)^{12} -
+                                \left( \frac{\sigma}{r - \Delta}
+                                \right)^{6} \right]; & r < (r_{\mathrm{cut}}) \\
+                             = & 0; & r \ge (r_{\mathrm{cut}}) \\
+        \end{eqnarray*}
+
+    See `Pair` for details on how forces are calculated and the
+    available energy shifting and smoothing modes.
+
+    Note:
+        To replicate the behavior of the SLJ potential in HOOMD v2, set
+        `hoomd.md.pair.Pair.r_cut` to ``r_cut_unshifted + delta``.
+
+    .. py:attribute:: params
+
+        The potential parameters. The dictionary has the following keys:
+
+        * ``epsilon`` (`float`, **required**) - energy parameter
+          :math:`\varepsilon` :math:`[\mathrm{energy}]`
+        * ``sigma`` (`float`, **required**) - particle size :math:`\sigma`
+          :math:`[\mathrm{length}]`
+        * ``delta`` (`float`, **required**) - radial shift
+          :math:`\Delta` :math:`[\mathrm{length}]`
+
+        Type: `TypeParameter` [`tuple` [``particle_type``, ``particle_type``],
+        `dict`]
+
+    Example::
+
+        nl = nlist.Cell()
+        expanded_lj = pair.ExpandedLJ(default_r_cut=3.0, nlist=nl)
+        expanded_lj.params[('A', 'A')] = dict(epsilon=1.0, sigma=1.0, delta=1.0)
+        expanded_lj.params[('A', 'B')] = dict(
+                                             epsilon=2.0,
+                                             sigma=1.0,
+                                             delta=0.75)
+        expanded_lj.params[('B', 'B')] = dict(epsilon=1.0, sigma=1.0, delta=0.5)
+
+    """
+    _cpp_class_name = 'PotentialPairExpandedLJ'
+
+    def __init__(self, nlist, default_r_cut=None, default_r_on=0., mode='none'):
+        super().__init__(nlist, default_r_cut, default_r_on, mode)
+        params = TypeParameter(
+            'params', 'particle_types',
+            TypeParameterDict(epsilon=float,
+                              sigma=float,
+                              delta=float,
+                              len_keys=2))
+        self._add_typeparam(params)
+        self.mode = mode
 
 
 class Yukawa(Pair):