+++ hoomd/pytest/test_simulation.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
+
 import hoomd
 import numpy as np
 import pytest
@@ -333,8 +336,9 @@ def test_run_limit(simulation_factory, lattice_snapshot_factory):
         sim.run(-1)
 
 
-def test_seed(simulation_factory, lattice_snapshot_factory):
-    sim = simulation_factory()
+def test_seed(device, lattice_snapshot_factory):
+
+    sim = hoomd.Simulation(device)
     assert sim.seed is None
 
     sim.seed = 42