+++ hoomd/write/table.py
@@ -1,6 +1,5 @@
-# Copyright (c) 2009-2021 The Regents of the University of Michigan
-# This file is part of the HOOMD-blue project, released under the BSD 3-Clause
-# License.
+# Copyright (c) 2009-2022 The Regents of the University of Michigan.
+# Part of HOOMD-blue, released under the BSD 3-Clause License.
 
 """Implement Table."""
 
@@ -313,8 +312,8 @@ class _TableInternal(_InternalAction):
             param_dict = ParameterDict()
             param_dict.update(state['_param_dict'])
             state['_param_dict'] = param_dict
-            del state['_param_dict']['output']
-            state['_param_dict']['output'] = None
+            state['_param_dict']._dict['output'] = None
+            state['_param_dict']['output']
             return state
         else:
             return super().__getstate__()