+++ CHANGELOG.rst
@@ -1,8 +1,36 @@
+.. Copyright (c) 2009-2022 The Regents of the University of Michigan.
+.. Part of HOOMD-blue, released under the BSD 3-Clause License.
+
 Change Log
 ==========
 
 v3.x
 ----
+
+v3.0.0-beta.13 (not yet released)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+*Added*
+
+* ``md.pair.ExpandedLJ`` - A Lennard-Jones potential where ``r`` is replaced with ``r-delta``.
+* Support nested modification of operation parameters.
+* ``wall`` - Define wall surfaces in the simulation box.
+* ``md.external.wall`` - Pair interactions between particles and wall surfaces.
+
+*Changed*
+
+*Fixed*
+
+* Compile error with ``Apple clang clang-1300.0.29.30``.
+* Incorrect OPLS dihedral forces when compiled with ``Apple clang clang-1300.0.29.30``.
+
+*Deprecated*
+
+* ``md.pair.SLJ`` - Replaced with ``md.pair.ExpandedLJ``.
+
+*Removed*
+
+* Leftover ``state`` logging category.
 
 v3.0.0-beta.12 (2021-12-14)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^